DEPREM_ANALIZI
Between 2020 and 2024, earthquakes in Turkey were analyzed, and the following conclusions were reached:

Number of Earthquakes: During this period, a total of X earthquakes of varying magnitudes occurred worldwide, with a significant number recorded in Turkey.

Average Magnitude: The average magnitude of the earthquakes was calculated as Y on the Richter scale. Notably, an increase in earthquakes with M ≥ 6.0 was observed, particularly in Turkey.